+++ app/cmd/commands/migrate.go
@@ -24,13 +24,17 @@ func NewMigrateRootCmd() *cobra.Command {
 				&models.Identity{},
 				&models.Credential{},
 				&models.RefreshToken{},
+
 				&models.Project{},
 				&models.Board{},
+				&models.BoardList{},
+				&models.Task{},
+				&models.Note{},
 			); err != nil {
 				log.Fatal(err)
 			}
 
-			log.Println("Database migrated")
+			log.Printf("Database migrated in %s\n", c.Config().Database.Url)
 		},
 	}