+++ app/models/identity.go
@@ -10,8 +10,8 @@ import (
 
 type Identity struct {
 	ID        uuid.UUID `gorm:"type:uuid;primaryKey" json:"id"`
-	FullName  string    `json:"full_name"`
-	Email     string    `gorm:"unique" json:"email"`
+	FullName  string    `gorm:"not null" json:"full_name"`
+	Email     string    `gorm:"not null;unique" json:"email"`
 	Active    bool      `json:"active"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`