+++ app/web/utils/page_models.go
@@ -1,0 +1,13 @@
+package utils
+
+type BasePage struct {
+	Title     string
+	User      *PageUser
+	CsrfToken string
+}
+
+type PageUser struct {
+	ID    string
+	Email string
+	Icon  string
+}